Refactor test models for ort_fusions (#2237)
Moving test models used for ort fusions to models folder for to avoid
bloating of root folder when more test models (for example,
whisper-decoder and whisper-encoder) are added to test models
