+++ onnxscript/rewriter/ort_fusions/models/_smollm_1.py
